Move atlas package into rove
