+++ pkg/rove/atlas.go
@@ -1,4 +1,4 @@
-package atlas
+package rove
 
 import (
 	"github.com/mdiluz/rove/pkg/maths"
