+++ pkg/rove/atlas_test.go
@@ -1,4 +1,4 @@
-package atlas
+package rove
 
 import (
 	"testing"
