+++ pkg/rove/chunkAtlas.go
@@ -1,4 +1,4 @@
-package atlas
+package rove
 
 import (
 	"log"
